# packages/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/buildings": {
      "post": {
        "operationId": "BuildingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuildingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The building has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new building",
        "tags": [
          "Buildings"
        ]
      },
      "get": {
        "operationId": "BuildingsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The buildings has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Building"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all buildings",
        "tags": [
          "Buildings"
        ]
      }
    },
    "/api/v1/buildings/{id}": {
      "get": {
        "operationId": "BuildingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The building has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "building not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a building by id",
        "tags": [
          "Buildings"
        ]
      },
      "patch": {
        "operationId": "BuildingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBuildingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The building has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "building not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a building by id",
        "tags": [
          "Buildings"
        ]
      },
      "delete": {
        "operationId": "BuildingsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The building has been successfully deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a building by id",
        "tags": [
          "Buildings"
        ]
      }
    },
    "/api/v1/buildings/{id}/restore": {
      "patch": {
        "operationId": "BuildingsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The building has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "building not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore a deleted building",
        "tags": [
          "Buildings"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "summary": "Login user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/check-auth-status": {
      "get": {
        "operationId": "AuthController_checkAuthStatus",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/seed": {
      "get": {
        "operationId": "SeedController_seedDB",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Seed"
        ]
      }
    },
    "/api/v1/common-areas": {
      "post": {
        "operationId": "CommonAreasController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommonAreaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CommonAreas"
        ]
      },
      "get": {
        "operationId": "CommonAreasController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CommonAreas"
        ]
      }
    },
    "/api/v1/common-areas/{id}": {
      "get": {
        "operationId": "CommonAreasController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CommonAreas"
        ]
      },
      "patch": {
        "operationId": "CommonAreasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommonAreaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CommonAreas"
        ]
      },
      "delete": {
        "operationId": "CommonAreasController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CommonAreas"
        ]
      }
    }
  },
  "info": {
    "title": "My Buildings API",
    "description": "Building Management API",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateBuildingDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main Building",
            "description": "Building name"
          },
          "description": {
            "type": "string",
            "example": "New building in the city",
            "description": "Building description"
          },
          "yearBuilt": {
            "type": "number",
            "example": 2020,
            "description": "Year built"
          },
          "propertyType": {
            "type": "string",
            "example": "RESIDENTIAL",
            "description": "Property type",
            "enum": [
              "RESIDENTIAL",
              "COMMERCIAL",
              "MIXED"
            ]
          },
          "address": {
            "type": "string",
            "example": "Main Street",
            "description": "Building address"
          },
          "district": {
            "type": "string",
            "example": "Main District",
            "description": "Building district"
          },
          "city": {
            "type": "string",
            "example": "Main City",
            "description": "Building city"
          },
          "province": {
            "type": "string",
            "example": "Main Province",
            "description": "Building province"
          },
          "postalCode": {
            "type": "string",
            "example": "15001",
            "description": "Building postal code"
          },
          "floors": {
            "type": "number",
            "example": 10,
            "description": "Number of floors"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+51 987 654 321",
            "description": "Building phone number"
          },
          "email": {
            "type": "string",
            "example": "contact@building.com",
            "description": "Building email"
          },
          "amenities": {
            "type": "array",
            "example": [
              "ELEVATOR",
              "PARKING"
            ],
            "description": "Building amenities",
            "items": {
              "type": "string",
              "enum": [
                "PARKING",
                "SECURITY_24_7",
                "ELEVATOR",
                "WHEELCHAIR_ACCESS",
                "WHEELCHAIR_LIFT",
                "FIRE_ALARM",
                "CAMERAS"
              ]
            }
          }
        },
        "required": [
          "name",
          "yearBuilt",
          "propertyType",
          "address",
          "district",
          "city",
          "province",
          "floors",
          "amenities"
        ]
      },
      "Building": {
        "type": "object",
        "properties": {}
      },
      "UpdateBuildingDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main Building",
            "description": "Building name"
          },
          "description": {
            "type": "string",
            "example": "New building in the city",
            "description": "Building description"
          },
          "yearBuilt": {
            "type": "number",
            "example": 2020,
            "description": "Year built"
          },
          "propertyType": {
            "type": "string",
            "example": "RESIDENTIAL",
            "description": "Property type",
            "enum": [
              "RESIDENTIAL",
              "COMMERCIAL",
              "MIXED"
            ]
          },
          "address": {
            "type": "string",
            "example": "Main Street",
            "description": "Building address"
          },
          "district": {
            "type": "string",
            "example": "Main District",
            "description": "Building district"
          },
          "city": {
            "type": "string",
            "example": "Main City",
            "description": "Building city"
          },
          "province": {
            "type": "string",
            "example": "Main Province",
            "description": "Building province"
          },
          "postalCode": {
            "type": "string",
            "example": "15001",
            "description": "Building postal code"
          },
          "floors": {
            "type": "number",
            "example": 10,
            "description": "Number of floors"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+51 987 654 321",
            "description": "Building phone number"
          },
          "email": {
            "type": "string",
            "example": "contact@building.com",
            "description": "Building email"
          },
          "amenities": {
            "type": "array",
            "example": [
              "ELEVATOR",
              "PARKING"
            ],
            "description": "Building amenities",
            "items": {
              "type": "string",
              "enum": [
                "PARKING",
                "SECURITY_24_7",
                "ELEVATOR",
                "WHEELCHAIR_ACCESS",
                "WHEELCHAIR_LIFT",
                "FIRE_ALARM",
                "CAMERAS"
              ]
            }
          }
        }
      },
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John Doe",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "password",
            "description": "User password"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john.doe@example.com"
          },
          "accessToken": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "accessToken"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "password",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateCommonAreaDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCommonAreaDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}