# packages/shared/src/types/prisma.types.ts
// This file was auto-generated by prisma-generator-typescript-interfaces

export type ReservationStatus = 'PENDING' | 'CONFIRMED' | 'CANCELLED'

export const ReservationStatusValues = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
} satisfies Record<string, ReservationStatus>

export type UnitType = 'APARTMENT' | 'OFFICE' | 'COMMERCIAL'

export const UnitTypeValues = {
  APARTMENT: 'APARTMENT',
  OFFICE: 'OFFICE',
  COMMERCIAL: 'COMMERCIAL',
} satisfies Record<string, UnitType>

export type UnitStatus = 'AVAILABLE' | 'OCCUPIED' | 'MAINTENANCE'

export const UnitStatusValues = {
  AVAILABLE: 'AVAILABLE',
  OCCUPIED: 'OCCUPIED',
  MAINTENANCE: 'MAINTENANCE',
} satisfies Record<string, UnitStatus>

export type PropertyType = 'RESIDENTIAL' | 'COMMERCIAL' | 'MIXED'

export const PropertyTypeValues = {
  RESIDENTIAL: 'RESIDENTIAL',
  COMMERCIAL: 'COMMERCIAL',
  MIXED: 'MIXED',
} satisfies Record<string, PropertyType>

export type UserRole = 'ADMIN' | 'MANAGER' | 'RESIDENT'

export const UserRoleValues = {
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  RESIDENT: 'RESIDENT',
} satisfies Record<string, UserRole>

export type Amenities =
  | 'PARKING'
  | 'SECURITY_24_7'
  | 'ELEVATOR'
  | 'WHEELCHAIR_ACCESS'
  | 'WHEELCHAIR_LIFT'
  | 'FIRE_ALARM'
  | 'CAMERAS'

export const AmenitiesValues = {
  PARKING: 'PARKING',
  SECURITY_24_7: 'SECURITY_24_7',
  ELEVATOR: 'ELEVATOR',
  WHEELCHAIR_ACCESS: 'WHEELCHAIR_ACCESS',
  WHEELCHAIR_LIFT: 'WHEELCHAIR_LIFT',
  FIRE_ALARM: 'FIRE_ALARM',
  CAMERAS: 'CAMERAS',
} satisfies Record<string, Amenities>

export type CommonAreas =
  | 'GYM'
  | 'POOL'
  | 'CLUB_HOUSE'
  | 'CAFETERIA'
  | 'EVENT_ROOM'
  | 'ROOF_TOP'
  | 'COWORKING_SPACE'

export const CommonAreasValues = {
  GYM: 'GYM',
  POOL: 'POOL',
  CLUB_HOUSE: 'CLUB_HOUSE',
  CAFETERIA: 'CAFETERIA',
  EVENT_ROOM: 'EVENT_ROOM',
  ROOF_TOP: 'ROOF_TOP',
  COWORKING_SPACE: 'COWORKING_SPACE',
} satisfies Record<string, CommonAreas>

export type Building = {
  id: number
  name: string
  description: string | null
  yearBuilt: number
  propertyType: PropertyType
  address: string
  district: string
  city: string
  province: string
  postalCode: string | null
  floors: number
  phoneNumber: string | null
  email: string | null
  isActive: boolean
  createdAt: Date
  updatedAt: Date
  deletedAt: Date | null
  manager?: User
  managerId: number
  commonAreas?: CommonArea[]
  amenities: Amenities[]
}

export type User = {
  id: number
  buildings?: Building[]
  firstName: string
  lastName: string
  email: string
  password: string
  role: UserRole
  createdAt: Date
  updatedAt: Date
  isActive: boolean
}

export type Unit = {
  id: number
  buildingId: number
  number: string
  floor: number
  type: UnitType
  status: UnitStatus
  createdAt: Date
  updatedAt: Date
}

export type CommonArea = {
  id: number
  buildingId: number
  building?: Building
  type: CommonAreas
  description: string | null
  capacity: number | null
  isActive: boolean
  createdAt: Date
  updatedAt: Date
  deletedAt: Date | null
  maxHoursPerReservation: number | null
  openTime: string
  closeTime: string
  daysAvailable: string[]
  reservations?: Reservation[]
}

export type Resident = {
  id: number
  firstName: string
  lastName: string
  phoneNumber: string | null
  email: string
  isActive: boolean
  createdAt: Date
  updatedAt: Date
  reservations?: Reservation[]
}

export type Reservation = {
  id: number
  commonAreaId: number
  commonArea?: CommonArea
  residentId: number
  resident?: Resident
  startTime: Date
  endTime: Date
  status: ReservationStatus
  notes: string | null
  attendees: number
  createdAt: Date
  updatedAt: Date
}